refactor constants template for recursion
Summary:
The current py3 constants setup heavily uses `types`, which is nice because much of the logic is type-specific. Unfortunately, it works poorly for recursive calls, as are needed for containers and struct migration to python. It ends up being less janky to add some logic to `mstch_py3` to expose the types where needed.

This diff takes the current `templates/python/types/constant_value.mustache` and adapts it to py3, so we aren't yet migrating the constants and structs from cbindings to python.

Reviewed By: createdbysk

Differential Revision: D66724441

fbshipit-source-id: cb822c3b1686ca374183f0221390c34d0754c29e