Fix expression tree nesting with await bug
Summary:
The lowerer introduces lambdas to allow lists of statements in expression trees, e.g. ``ExampleDsl`{$x = 1; $y = 2;}` `` lowers to ``ExampleDsl`(() ==> {$x = 1; $y = 2;})()` `` before desugaring. This is an expression tree lambda, not a Hack lambda, and thus processing it shouldn't interfere with the code looking for awaits and yields should they appear in splices, nor should we forget if we are in an enclosing async Hack function. However, the call to `reset_for_function_body` was doing exactly that.

Improve error message if a yield appears in an expression tree

Reviewed By: andrewjkennedy

Differential Revision: D66822415

fbshipit-source-id: 1b89af3ab838473c51d641331fdce73bdab010a4