Errors streaming: fix bug where --error-format raw does now print "No errors!"
Summary:
`Errors.format_summary` (the function printing "x errors, y warnings" or "No errors!" at the end of printing errors) first matches on format and if the format is 'plain' or 'raw', it does not print anything. However, `format_summary` is always called like this whenever it's called (more or less):
```
if List.is_empty errors then
  Printf.printf "No errors!"
else
  Errors.format_summary ...
```
Except when streaming errors.
This means that `hh --error-format raw` will fail to print "No errors!" when there are no errors when error streaming is on.

I could replicate the pattern and do `if List.is_empty errors then print_no_errors () else Errors.format_summary ...` in the error streaming code, or I could do what this diff does: change `format_summary` so that when error format is raw or plain, it either prints "No errors!" if there are no errors or it prints nothing otherwise. Because of the way `format_summary` is called, the code from this diff will only be hit for error streaming, which limits the impact of this diff.

Reviewed By: enetsee

Differential Revision: D66961555

fbshipit-source-id: c9d585d72f0997b1bf1461ee4fdebe156b2465b5