Produce reason information that explains provenance of like type even through SDT calls
Summary: The provenance of like-types gets lost across SDT calls. For example, if `~C` originates in a call to an abstract method (which has a pessimised return type) and then that value gets passed to an SDT function that returns a `D`, then the dynamic gets propagated but the provenance in `~C` is lost. Let's improve this.

Differential Revision: D66876596

fbshipit-source-id: 86335e30be595b8690bbfd30fedae5cfb144e3d0