Always generate abstract types.
Summary:
# What?

To prepare for the general availability (GA) of abstract types, unconditionally generate `thrift_abstract_types.py`.

NOTE: The `experimental_generate_abstract_types` flag still gates the use of the abstract types. Given that the GA is close, leave the flag with a name (`generate` rather that `use`) that doesn't completely match it's behavior.

Reviewed By: yoney

Differential Revision: D66942161

fbshipit-source-id: 7ca93ce2e176ec0c182e7dcc41f6002a4550e832