Improve robustness of test/zend/good/ext/standard/tests/strings/crypt.php
Summary:
When no salt (2nd arg) is passed to crypt(), it uses C++ std::rand() to produce a hash,
causing the result to change based on whether calls to std::rand() are
inserted/removed prior to the call to crypt().  Change the test to match
any string hash returned by crypt() in such cases.

Reviewed By: mdko

Differential Revision: D66978809

fbshipit-source-id: f586ded68861f5b46aac23a87f95b481401c4e46