Improve crash error messge for ThreadManagerLoggingWrapper when run with Resource Pools
Summary:
# Issue
The investigation of S474343 was long due to confusing error indicating that `primaryThreadManager_` is null, even though it was initialy setup by the service, but further overriden by ResourcePools migration.

# Change
Improve error message during crash, so that it's clear that this is caused by call to deprecated ThreadManager API after Resource Pools migratioin.

Reviewed By: sazonovkirill

Differential Revision: D66910562

fbshipit-source-id: acd3c3b3ac99bb99fbcd80e991da9ba6a598e670