Enable abstract types in all thrift-python fixtures.
Summary: To ensure that abstract types enable correctly when they are generally available, add the `experimental_generate_abstract_types` flag to all thrift-python fixtures that don't have that flag. In the diff where abstract types become GA, except for a few select cases, which I will document in that diff, the difference should be the removal of the `Experimental` comment.

Reviewed By: yoney

Differential Revision: D66944942

fbshipit-source-id: 8c8981194c01d8e701860f0d0f1a3f887835fca6