Add Gatekeeper for creating a case in SF through thrift service
Summary:
We are using thrift service for providing our case creation API from Care Platform to BSE team in fbcode.
They will call a python method, that calls our case creation method in www and all of that is implemented with thrift service.

In this diff I have added a Gatekeeper in order to enable the feature only for developers while the feature is in development.

Differential Revision: D66708306

fbshipit-source-id: 25453a22b452b7c4ee1f95dfee397aed4f279da4