fix enums -> metadata -> types -> enums circular dependency
Summary:
Enums are terminal types, so the `thrift_enums` module should have no dependency, direct or indirect, on `.thrift_types`.

This is currently not the case. We have a dependency path `thrift_enums` -> `thrift_metadata` -> `thrift_types`.

This is a problem because `thrift_types` also depends on `thrift_enums`, completing a circular dependency cycle, making import behavior dependent on import order.

This is also causing problems for thrift-py3 because the py3 enums module (`types_impl...`) is unnecessarily depending on `.thrift_types` modules of all recursive dependencies.

The fix is to move the enum metadata generation from `thrift_metadata` to `thrift_enums`, making `thrift_enums` a leaf in the build graph.

Reviewed By: yoney

Differential Revision: D66973310

fbshipit-source-id: c4214a009fde934c3615a7361acb1cc81bf2dbc6