First-class polymorphic function pointer for top-level functions
Summary:
Modifies typing of function pointers for top-level definitions so that they have universally quantified types and makes the corresponding changes to subtyping to allow these types in subtype position.

For polymorphic function types in subtype position we just instantiate with fresh, invariant type variables (to avoid solving) then solve the proposition between the supertype and the instantiated function type.

For polymorphic function types in supertype position we bind each quantifier in the type parameter environment, solve the proposition in the updated context then restore the context i.e. unbind the quantifiers. Because we use a fixed, leftmost, eager depth first search strategy we are guaranteed that the context is correct for all subgoals.

Reviewed By: andrewjkennedy

Differential Revision: D66543489

fbshipit-source-id: 55a2f4b51304195b2c0f88027a9a96c9f705092a