fix filtering of init warnings when on public commit
Summary:
We have logic to filter out warnings resulting from the first typecheck after init, if we are on a public commit. However we apply that filtering at the end of typechecking. This is incompatible with error streaming where errors are streamed to client as soon as they are found.

This diff fixes that by doing the filtering just before streaming the errors. This requires threading down the boolean corresponding to "is init typecheck and is public commit", and bubbling up the resulting set of warning hashes.

Reviewed By: madgen

Differential Revision: D67035357

fbshipit-source-id: 1603e26eeed9db63a00c8d559db16b41c37a6ebf