use a hazard pointer to make updates safe
Summary:
Previously sink and stream had a reference to the payload serializer that could have pointed to a version of the payload serializer that could be deleted.

Uses a hazard pointer so that is safe to delete the PayloadSerializer while other threads are using it. Hazard pointer will clean up the object after its done being used.

#build_more

Reviewed By: iahs

Differential Revision: D66941128

fbshipit-source-id: 98606f0250d7967d61b7b85cc688b1c770dd4ccf