A better Hg.Rev.equal
Summary:
I will use this in the next diff.

We have the abstract type `Hg.Rev.t` representing an hg revision. Whether two `Hg.Rev.t` are `equal` should have the same meaning as whether two hg revisions are the same, e.g. revs "5aa0fe85e2" and "5aa0fe85e22dc7f8e67439b9c44b9b6f3c31806a" are the same revs.

`Hg.Rev.t` is a string under the hood, but equality of `Hg.Rev.t` is different from string equality. Instead, it is whether one rev string is a prefix of the other. This diff implements that.

The only current usage or `Hg.Rev.equal` (in watchman.ml) unfortunately relied on it being implemented as `String.equal`, because it was occasionally comparing an empty-string rev to a non-empty-string rev. To fix that, we use a new constructor `Hg.Rev.of_string_check_empty` which returns `None` if passed an empty string.

Differential Revision: D67142116

fbshipit-source-id: 0143cde766b301a9cf7659008957ae7e2daa75d3