Introduce `expression` AST node
Summary:
Interpolations and conditions are permitted to contain arbitrary expressions, not just variable lookups. Of the 3 types of expressions in the spec only one is currently implemented, but we want to avoid hardcoding that fact to allow more types to be implemented.
We intentionally don't support expressions as arguments to section blocks as those are deprecated.

Reviewed By: praihan

Differential Revision: D66901989

fbshipit-source-id: cd5b3b92627049ac37b05eb8869af18d14a221fd