remove thrift_enums.pyi
Summary:
We have `thrift_types.pyi` because the struct meta-classes are way too complicated for pyre to handle, so we have to manually specify the struct interfaces.

However, we don't have this issue with enums. By specifying that `thrift.python.types.EnumMeta` extends `enum.EnumMeta`, in `thrift.python.types.pyi`, then we enable the type system to understand that are enums are enums.

All that remains is fixing some missing annotations in `.thrift_enums.py`, and then we can remove `.thrift_enums.pyi`.

Reviewed By: yoney, Filip-F

Differential Revision: D67256957

fbshipit-source-id: fbc56134f0043c4ccd5e6c340a8ca49033e20eda