Warnings: always discard init warnings when saved state distance is non null
Summary:
Context: We want to only show warnings caused by user changes. We do this by storing all warnings in the saved state, and hiding warnings from that saved state. If the mergebase does not have a saved state however, that means some unrelated warnings introduced between the saved state rev and the mergebase may be shown. So far we've mitigated this by hiding all warnings if hack is initialized from a public revision. However this still shows warnings when hack is initialized from a private commit.

To keep warnings high-signal, we decided we'd rather hide too many warnings than show too many, so this diff is more aggressive about hiding warnings. Instead of hiding warnings if initialized on a public rev, we hide all warnings unless the mergebase has a saved state.

Reviewed By: andrewjkennedy

Differential Revision: D67140009

fbshipit-source-id: 838ff43ea7b92c3b53f19fa1257f0a38b087c6b7