Delete services reflection in windtunnel/cogwheel/fuzzing
Summary:
There are several files here using services reflection to accomplish the goal of letting you send fuzz data at a running service. This is easily the most complex use of services reflection anywhere, as the fuzzing service has to be able to understand what kind of data a service is supposed to be receiving and how to initialize it.

This required a whole lot of janky workarounds to get functioning with py3 metadata (as py3 reflection gives you access to the actual types, while metadata doesn't).

Chief among the janky workarounds is `_fbthrift_annotations_DO_NOT_USE`, which is necessary because the service introspection requires access to the full modules being passed to a function. This is accomplished in thrift python by using `[service function].__annotations__`, which gives you all the type annotations on the given service method. This is a horrific solution that relies on type annotations for runtime correctness, but unfortunately we need to make this function in thrift py3, and cython methods don't have any equivalent to `__annotations__`, so as a painful workaround I had to add `_fbthrift_annotations_DO_NOT_USE_{{function:name}}` that holds this annotations data for thrift py3 clients/services.

Also I needed to specifically populate this data with period separated module paths (ie. `some.import_path.types` vs `some_import_path_types`) because the solution that the cogwheel people implemented was just reversing our name mangling, and for thrift python this works sort of (we replace `.` with `__`, which isn't unique but is at least less common, in thrift py3 we replace `.` with `_` which is very common), but for thrift py3 it breaks. Fixed this by just giving cogwheel the actual import path in the awful annotation dict I'm adding to the clients/services.

I don't think this will add much build time as this is just a pre populated dict of strings. If anyone is aware of a better way of accomplishing this though I am open.

Reviewed By: ahilger

Differential Revision: D66849970

fbshipit-source-id: 95a7f609436f59ca3c303770be6f5e66ca5a5918