Delete thrift py3 services reflection
Summary:
The ability to reflect on services is something that is no longer used, so deleting the relevant code that exposes it.

There are some SvIf bindings in `services_reflection.pxd`, and those need to remain because they're needed in the services/clients files (I tried moving them into metadata but that broke for files that define types but not services, as you get a py3-types -> py3-metadata -> cpp2-services dependency chain), so renaming those files to services_interface to reflect what they actually mean and deleting `services_reflection.py`.

#buildall

Reviewed By: ahilger

Differential Revision: D67328763

fbshipit-source-id: 332a2d4b8c8c25364b38d4c304c2b51b33f3345e