Reland D67328763: Delete thrift py3 services reflection
Summary:
D67328763 got reverted because there was an IG use case for services reflection that was just calling the internal `__get_reflection__` method and using pyre-ignore, so we had no way of catching it (and sandcastle never caught it too). Fixed that in the previous diff in this stack so relanding.

Original Description:

The ability to reflect on services is something that is no longer used, so deleting the relevant code that exposes it.

There are some SvIf bindings in `services_reflection.pxd`, and those need to remain because they're needed in the services/clients files (I tried moving them into metadata but that broke for files that define types but not services, as you get a py3-types -> py3-metadata -> cpp2-services dependency chain), so renaming those files to services_interface to reflect what they actually mean and deleting `services_reflection.py`.

Reviewed By: yoney

Differential Revision: D67456043

fbshipit-source-id: 62626fc2e298049ed523852ff0e7008ee35c61d4