hackrs: print singleton unions and intersection in a parsable way
Summary:
The type printer in hackrs currently prints singleton unions as `(|T)` instead of just `T`. The reason is explained in [this comment](https://www.internalfb.com/diff/D52491369?transaction_fbid=3812697109009854): the code was ported from the ocaml printer `Typing_print.Full`, which is also used for debugging. However, in rust, if one wants to debug, they can just use the `Debug` trait instead of the `Display` trait.

Furthermore, the printer code is used in ext-decl and the output ends up being re-parsed later. `(|T)` fails parsing, so I'm changing the output to be just `T`. [This](https://fburl.com/code/rb0mzmh3) is one place where the output of the printer is used to populate the `type_` fields of an ExtDeclTypeDef. And this later flows into parsing code like [this](https://fburl.com/code/5bogu3ui).

Reviewed By: paulbiss

Differential Revision: D67335103

fbshipit-source-id: 0a12ea2c851ec44c357919d377be46bbbe5cff58