core implementation of experimental `require this as C` trait constraints
Summary:
We introduce the experimental `require this as C;` trait  constraints.  While `require class C;` trait constraint add the `this = C` constraint to the environment when typechecking the trait methods (and thus require `C` to be `final` for soundness), `require this as C;` add the `this <: C` constraint and don't require `C` to be final.  Contrarily to `require extends`, `this` does not need to be a strict subclass of `C`.

In the long term we might rename `require class C;` into `require this = C;` for uniformity.

The implementation reuses the `require class` search for elements, and differs from the `require class` one only in the following aspects:

- `this <: C` is added to the environment instead of `this = C`
- the check for `C` being final is removed.

Feature is gated by `<<file:__EnableUnstableFeatures('require_constraint')>>`

Reviewed By: andrewjkennedy

Differential Revision: D66183523

fbshipit-source-id: 2518380debad2215afdf441fce4b7453e11edd2f