Add single-line pragma
Summary:
We have a number of partials that render fragments of lines, for example the template argument of the return type of a method. In order to keep the output readable we suppress newlines in these partials by commenting them out, which adds O(n) syntactic noise. This hurts readability and adds overhead to modifying these partials.

Handlebars supports [whitespace control](https://handlebarsjs.com/guide/expressions.html#whitespace-control) for statements, which is less disruptive than our comment-based approach but still introduces O(n) syntactic noise.

Instead we introduce an O(1) syntax for suppressing all newlines in the current partial. To prevent confusing action-at-a-distance it is not applied transitively to included partials, and the included partial is also expected to have this pragma already.

Reviewed By: yoney

Differential Revision: D67611032

fbshipit-source-id: d8486196a98763483e3a47e5774f0ea8b0d04c1b