Implement parsing of user-defined functions
Summary:
This diff implements only the parsing part of user-defined functions. The renderer currently rejects them and support for them will be added in a future diff.

The change here adds in a new kind of `function_call` in the AST called `user_defined`. User-defined functions are accessed via a `variable_lookup`.

Additionally, names are changed to match nomenclature used in the spec (builtin), and error messages have been updated to be more consistent with others.

Reviewed By: createdbysk

Differential Revision: D67587583

fbshipit-source-id: 12ba36ff8153b2bf751ea4f0568cd2b6f3aba8bc