Treat compiler options as immutable after initialization.
Summary:
# What?
Treat the compiler options as immutable after their `t_hack_generator::process_options` initializes them.

# Why?
There are some cases in render const value helper where the hack generator forces the use of arrays. It currently temporarily modifies the `arrays_` compiler to achieve this outcome. When a subsequent change removes the `arrays_` option and relies on a combination of `no_use_hack_collections_` and a new `hack_collections_` option, there will be no `arrays_` option to modify, and there is no easy mechanism to correctly modify the state of the  compiler options to achieve the same outcome. In addition, this modification of the compiler options makes it difficult to reason about the state of the generator.

# What to do?
- Pass a parameter that forces the use of arrays.
- Delete the code that modifies the value of the `arrays_` compiler option.

# Improvement
To prevent issues with accidental correctness in spite of incorrectly passed parameters, change optional parameters to required parameters.

# Fixtures
Generated fixtures must not change.

Reviewed By: yfeldblum, yoney, rmakheja

Differential Revision: D67569500

fbshipit-source-id: 63d2a29539640d7e7de4af4a7ce421b0156868ac