Implement user-defined functions (native_function)
Summary:
This diff implementation the lowest-level API for user-defined functions: `native_function`.

The implementation here matches the existing whisker spec. A lot of the details are covered in doc blocks within the code.

A future diff / project will build a DSL on top of `native_function` to make it more ergonomic / declarative when authoring in C++.

The parser was built in {D67587583} and literals in {D67608814}.

Reviewed By: iahs

Differential Revision: D67617014

fbshipit-source-id: 993bb5781dfa9293b65755e03d6d75642521abbd