Use single Rust staticlib for FFI
Currently, each Rust FFI library is built and linked into the HHVM
binaries as separate static libraries. This is [explicitly
unsupported](https://cxx.rs/build/other.html#linking-the-c-and-rust-together)
by the `cxx` crate, which recommends either using rustc as the final
linker or using a single Rust staticlib for the Rust components, and has
been the cause of linking errors for the last couple of years (see
T146965521).

The former does not seem to be feasible for HHVM OSS, so go with the
latter approach and combine Rust FFI libraries into a single staticlib,
then link that into HHVM. This will likely need corresponding changes in
the internal buck2 build.