Implement source backtraces on name lookup error
Summary:
This diff adds a "stack trace" when a name lookup error happens. This can be expanded to other errors in the future.

This will show the chain of partials that led to an error which can greatly improve debugging experience.

On failure, the output might be something like:
```
The source backtrace is:
#0 partial-3 <line:1, col:3>
#1 partial-2 <line:1, col:5>
#2 partial-1 <line:2, col:3>
#3 path/to/test.whisker <line:2, col:1>
```

This change also incorporates the code for `{{#pragma single-line}}` into it because the state that needs to be tracked is largely the same.

Reviewed By: iahs

Differential Revision: D67629403

fbshipit-source-id: 64f7fb1495bbfb5fc12da11f49141eeed1959864