Make pragma single-line → ignore-newlines
Summary:
`single-line` is not exactly correct because a `single-line` file can partially apply another non-`single-line` file and end up with multiple lines. Hence, I've changed the name to `ignore-newlines` which is slightly clearer and matches the codebase.

Also updated the Whisker spec and `print_ast` output to be more consistent with others.

Reviewed By: iahs, createdbysk

Differential Revision: D67637625

fbshipit-source-id: 2ca22a57b1df6c63279bd943188aa604ed5e18f7