Update and fix bundled boost build
Since Boost is a C++ library dependency, we should force building its
bundled version when building against libc++.

Take this opportunity to update to a more recent Boost version, and use
B2 options[1] to set the C++ version and toolchain from the
superproject. Remove an outdated macOS patch that no longer applies.

Add the chrono and date_time components to the list of components to
build as they're needed by mvfst.

[1] https://www.boost.org/doc/libs/1_87_0/tools/build/doc/html/index.html