Support schema registration after merged schema is accessed
Summary:
With python's lazy imports and dynamic linkage's deferred loading we can have situations where not all schemas are registered at the time of load. Instead of throwing in that scenario, update the merged schema with the new schema.

The synchronization scheme here prioritizes simplicity over performance, as the eventual consumption of the merged data will likely take a different form based on the runtime representation we choose to expose to users and optimizations made now may not be applicable later.

The SharedMutex here will be uncontended during the majority of registration happening before main() and we will only make copies of the merged data in cases where there is an active reference which will also not be the case before main.

Reviewed By: yoney, praihan, pranavtbhat

Differential Revision: D67677464

fbshipit-source-id: bff56ef632a9e289a17e343a25e812224ad0ce33