Provide less cryptic error for missing systemlib sections
If a new extension systemlib isn't wired up with the build system, its
unit emitters and decls won't be embedded in the HHVM binary, causing
a cryptic crash when trying to run Hack code ("Invalid varint value: too
few bytes."). One then has to fire up a debugger to determine what systemlib is
missing.

Instead, raise a more informative error if a specific systemlib was not
found in the binary.