Remove arraysets if arrays present
Summary:
# What?
Remove `arraysets` if `arrays` present.

# Why?
This combination can lead to runtime failures.

# Fixtures
Generated fixtures must not change.

# Context
`arraysets`, `no_use_hack_collections`, `stricttypes`, `array_migration`, `shape_arraykeys`, `const_collections` are all compiler options that control the generation of container fields. To make code generator simpler and easier to reason about, identify which of these can be removed/merged.

The new options based on the ones that are currently in use:

`legacy_arrays_` replaces `no_use_hack_collections_`.
`hack_collections_`, which was the implicit default in the absence of `arrays_` and `no_use_hack_collections_`, is now explicit.
`arrays_` will eventually become the default and cease to exist as an explicit option.

# The steps
The item in bold corresponds to the current diff.
1. Use new compiler options based on the ones that already exist.
1. Make arrays the default option.
1. Use the legacy arrays + hack collections logical equivalent of arrays.
1. Introduce hack collections wherever necessary.
1. Replace no use hack collections with legacy arrays.
1. **Remove arraysets if arrays present.**
1. Rename `no_use_hack_collections` compiler option to `legacy_arrays`.
1. Add `hack_collections` compiler option.
1. Remove references to the `ARRAYS` compiler option.
1. Delete the `arrays` compiler option.

Reviewed By: rmakheja

Differential Revision: D67573143

fbshipit-source-id: cad66b814fdd3f81c2862ce11d41cc0d0d07310a