recursive case type: make bound check less restrictive
Summary:
This diff aims at supporting cases like:
```
case type CT1 as shape(?'a' => nonnull) = shape(?'a' => CT1);
case type CT2 as shape(?'a' => shape(?'a' => nonnull)) = shape(?'a' => CT2);
```
The previous behaviour was to localize CT1/CT2 as mixed and therefore fail the bound check.
That localization to mixed was due to passing `~report_cycle:CT1/2` to `localize`. Upon detection of any cycle, localization stops and creates a mixed. Passing `~report_cycle:CT1/2` makes it localize the CT1/2 to mixed, which is wrong for further usages.

This diff calls localize twice, once to report cycles, and once to use the result for bound checks and other checks.

Reviewed By: andrewjkennedy

Differential Revision: D67938199

fbshipit-source-id: bdd850d909294abafaba5ff76fcc37f1229515e3