Delete arrays as a compiler option.
Summary:
#What?
Delete `arrays` as a compiler option.

#Why?
`arrays` are now the default in the absence of `legacy_arrays` and `hack_collections`, and cease to exist as an explicit option.

# Fixtures
- Delete `arrays` from `cmd` files.
- Generated fixtures must not change.

# Context
`arraysets`, `no_use_hack_collections`, `stricttypes`, `array_migration`, `shape_arraykeys`, `const_collections` are all compiler options that control the generation of container fields. To make code generator simpler and easier to reason about, identify which of these can be removed/merged.

The new options based on the ones that are currently in use:

`legacy_arrays` replaces `no_use_hack_collections`.
`hack_collections`, which was the implicit default in the absence of `arrays` and `no_use_hack_collections`, is now explicit.
`arrays` are now the default in the absence of `legacy_arrays` and `hack_collections`, and cease to exist as an explicit option.

# The steps
The item in bold corresponds to the current diff.
1. Use new compiler options based on the ones that already exist.
1. Make arrays the default option.
1. Use the legacy arrays + hack collections logical equivalent of arrays.
1. Introduce hack collections wherever necessary.
1. Replace no use hack collections with legacy arrays.
1. Remove arraysets if arrays present.
1. Rename `no_use_hack_collections` compiler option to `legacy_arrays`.
1. Add `hack_collections` compiler option.
1. Remove references to the `arrays` compiler option.
1. **Delete the `arrays` compiler option.**

Reviewed By: rmakheja

Differential Revision: D67618067

fbshipit-source-id: d6c20bd9343f868b1492497f8e41b72ec91c45a6