Allow user-defined functions to abstract over array-like and map-like native_objects (positional arguments)
Summary:
There are two ways each (in practice) to have arrays and maps in Whisker:
* `whisker::array` and `mstch_array` (wrapping `mstch::array`)
* `whisker::map` and `mstch_map` (wrapping `mstch::map`)

This can get really hard to deal with in user-defined `native_function` and can be error prone. This diff changes `context::argument<T>` such that accessing `array` or `map` forces the user to interact with `array_like` and `map_like` so that `native_function` implementations are able to support all the cases above seamlessly.

This diff:
* Introduces `array_like` and `map_like` types.
* Adds support for the above for *positional* arguments

Named arguments to follow in a future diff.

Reviewed By: vitaut

Differential Revision: D67920648

fbshipit-source-id: c4d1cbbb04443a5e0a5571fd1cd55423810d491d