Make "this" the implicit context object
Summary:
Today `.` is used as the implicit context object. But there's a problem...

An expression like `(func .)` has the intent of calling `func` with the value of the implicit context object. However, the parser thinks that `.` is a separator for variable lookup. The grammar is just ambiguous here.

The fix is to take a keyword we've already reserved, `this`, and make it also refer to the implicit context object.

Reviewed By: yoney

Differential Revision: D68005178

fbshipit-source-id: d34cf65e350ca70ae6dc061bda7c033bc041d78b