refactor make_instance for const& compat
Summary:
Cython is pretty dumb, so it doesn't understand that it can convert rvalue function returns to const& params. On top of that, it assigns (in generated .cpp) the result of every function call (in .pyx) to a local variable that has to be default constructible. That means we can't do something clevel like const casting to fix the first issue.

This specifically affects `set::insert(const T&)` and `map` key, where we use `map::operator[](const T&)`.

This is a pre-requisite to cleaning up converter logic and removing transitive includes.

V2: # / buildall

Reviewed By: yoney

Differential Revision: D68124625

fbshipit-source-id: 8bc5325d0320a832dc596abe6feb254710ee3619