Expose setting socket queue timeout to Python
Summary:
When migrating to thrift-python, we saw S480390. The fix normally would involve setting socket queue timeout, but it is not possible in the python implementation.
Add methods necessary to support this.

Note: I chose to change the arg type of the setter method in C++ rather than using std::chrono::nanoseconds in cython to make the non-observer methods consistent between get and set as well as consistency across the py/cpp bridge methods. Also it helps that all users were already passing milliseconds and relying on implicit conversion (TIL that it works milli -> nano but not nano->milli; makes sense tho)

Reviewed By: Filip-F

Differential Revision: D68093455

fbshipit-source-id: bfab83904f7700a3da447ddc2b580b20258f81fa