Type splat well-formedness: must be a tuple
Summary: The type splat operator `...t` should be used only on something known to be a tuple type, typically a generic with a bound e.g. `T as (int, nonnull...)`. Currently we don't enforce this. By analogy with generic types that have constraints (e.g. `keyset`), we should check in `Typing_type_wellformedness`. It's enough to check that the splat type is a subtype of `(mixed...)`, the top tuple type.

Reviewed By: francesco-zappa-nardelli

Differential Revision: D68273829

fbshipit-source-id: 2a63a1c38062577454f068a2fb86596c6e8c3758