Add call+fallback for streamJournalChanged to watchman
Summary:
# Context

SubscribeStreamTemporary is a function that allows users to recieve notifications whenever there is a change to the edenFS journal. Despite intended to be a temporary function, it's been named as such for about 6 years.

# This Diff
This diff replaces the call for watchman to subscribeStreamTemporary with a call for streamJournalChanged. It adds a fallback call for older versions without streamJournalChanged

Reviewed By: jdelliot

Differential Revision: D68240121

fbshipit-source-id: 0c3f9e98060bf80c1e0e78b06d8ac102a8a19060