add counters for memcache/tao latency
Summary:
This replaces a bunch of dyno counters which are being phased out.

All it does is take the values from server notes and put them in a counter. It's not the best way of doing things, but I don't currently have the wherewithal to plumb this into the corresponding extensions.

I will follow up later with thrift and mysql.

Reviewed By: AhmedAbdellah19

Differential Revision: D65980932

fbshipit-source-id: f7eb2d4ae32ddb26affcbf78a3001274cade4ba4