Implement LSP method $/setTrace
Summary:
We already have `$/setTraceNotification` implemented, but it turns out nowadays (or sometimes?), `$/setTrace` is received. The LSP doc only mentions `$/setTrace`, and `$/setTraceNotification` is [apparently](https://github.com/microsoft/language-server-protocol/issues/109) VS Code specific, but I couldn't find any info about the difference or history.

So I'm adding an implementation for `$/setTrace` and make it the same as the implementation for `$/setTraceNotification`.

Reviewed By: mheiber

Differential Revision: D68266300

fbshipit-source-id: 142b39e0943d9192061e796456e287cf791ce804