Bug fixes for adt data constructors in GADTs
Summary:
Fixes bugs for GADTs

OCaml GADTs can declare ADT-style data constructors, e.g.

```
type _ t =
  | GADT_data_ctor : 'a t
  | ADT_data_ctor
```

ppx_transform doesn't currently handle this case and fails with an exception. It is also using the wrong test to determine is a variant is a GADT - it currently looks if the first constructor has a 'return type' but it should be looking if any data constructor has a return type.

This also fixes a bug whereby phantom types are not distinguished from types occuring anywhere in the types for which we are deriving the transform

Reviewed By: mheiber

Differential Revision: D68214584

fbshipit-source-id: f871bb9e7f09c5d9a470b4dbb67aa0deab151eed