Move like type stripping to ppx_transform
Summary:
We should be restricting the use of the hand-written visitor for ty_ in favor of the function derived using ppx_transform with a view to deprecating that visitor.

I've added a [transform.explicit] annotation to [Tunion] to make this pass as simple as possible but you could of course do this by deconstructing the [locl_ty] as part of the pass.

Reviewed By: mheiber

Differential Revision: D68413396

fbshipit-source-id: ee658a2c43c3810624cc0e8b414edef49da0e3a8