Invalidate all folded decls upon file change on disk
Summary:
Last year, Lucian has tried a lot of different experiments to improve the overall correctness/performance trade-off, attempting to improve one while not damaging the other. After discussion with him, we agreed the change in this diff is a reasonable and simple adjustment to the toggle in favour of correctness and to the expense of performance.

The reason this is needed is to fix user-reported discrepancies between the IDE and the CLI.

What we do in this diff: upon file save, we invalidate all folded decls.
What we did before: we would invalidate only the folded decls of the symbols in the file being saved.

Argument for why this is acceptable in terms of performance: what is costly is parsing the shallow decls, not actually folding the decls. We keep invalidating only the necessary folded decls.

Reviewed By: madgen

Differential Revision: D68018098

fbshipit-source-id: dc702cd55314a3523bf52572bd7b63ce388c1ec7