Add set_rdonly function
Summary: This diff adds a new function to the `Options` struct in the `lmdb.rs` file. The `set_rdonly` function allows the environment to be opened in read-only mode, preventing any write operations from being allowed. This is useful in situations where the database is being used for read-only purposes, such as when serving static content. The function sets the `MDB_RDONLY` flag to enable read-only mode.

Reviewed By: nt591

Differential Revision: D68428162

fbshipit-source-id: 67d91d4162af4e33639b1c03382c334ebaabef0b