Make NoAutoLikes automatically inherited when overriding
Summary:
Marking a method `<<__NoAutoLikes>>` causes the return type (e.g. `vec<int>`) to be interpreted literally i.e. without any implicit like-type. Overriding methods that have the same return type, but for which a like-type is added implicitly (e.g. `~vec<int>`) will inevitably be rejected, because `~t` is not generally a subtype of `t`. So we may as well "inherit" the `<<__NoAutoLikes>>` attribute automatically, and save a ton of annotations.

Note that this just applies to ordinary inheritance: if any interface method is declared `<<__NoAutoLikes>>`, implementing methods will still need to be marked the same.

Differential Revision: D66594684

fbshipit-source-id: 72437864fef9990b979f60600c44567cdea8b61f