Refactor internal code action repr
Summary:
The current encoding which uses a combination of constraints, polymorphic variants and GADTs doesn't seem to be buying us a lot and makes adding additional 'things' which will become code actions a little tricky.

This diff backs out the more complex encoding in favor of plain old ADTs and a factored common type.

Reviewed By: mheiber

Differential Revision: D68443044

fbshipit-source-id: c44ec8cb9f077e6972ffd6fdad2de5725f48ef4f