Open in CodeHub API Docs
Summary:
This diff and the next add a menu item "Open in CodeHub API Docs" in the editor context menu, right under "Open in CodeHub".

This relies on a custom LSP method "custom/TopLevelDefNameAtPos", which returns the name of the top-level definition containing the current cursor. This itself is implemented using a tweaked version of IdentifySymbolService which is made declaration-span aware.

Reviewed By: mheiber

Differential Revision: D68416611

fbshipit-source-id: 7557bbde6f9e7ad0ea0788de67c5484d5980c84f