JSON representation of open tuple types
Summary:
Add printing and parsing for JSON representation of tuple types with optional, variadic, and splat fields.

Representation is similar to that used internally. A tuple type `(int, string, optional bool, nonnull...)` would be represented
```
{
    "src_pos": ...
    "kind": "tuple",
    "is_array": false,
    "args": [
      { "src_pos": ..., "kind": "primitive", "name": "int" },
      { "src_pos": ..., "kind": "primitive", "name": "string" },
    ],
    "optional_args": [
      { "src_pos": ..., "kind": "primitive", "name": "bool" }
    ],
    "variadic_arg": { "src_pos": ..., "kind": "nonnull" }
}
```
The `optional_args` field is omitted if there are no optional fields, likewise `variadic_arg` is omitted if the variadic is omitted (i.e. is `nothing`).

A tuple type `(int, ...T)` containing a type splat would be represented
```
{
    "src_pos": ...
    "kind": "tuple",
    "is_array": false,
    "args": [
      { "src_pos": ..., "kind": "primitive", "name": "int" },
    ],
    "splat_arg": { "src_pos": ..., "kind": "generic", "is_array": true, "name": T, "args": [] }
}
```

Reviewed By: viratyosin

Differential Revision: D68491256

fbshipit-source-id: d28dc4aebfccb5fe86836747d508de1a4ee212ab