do not overestimate numMethods in presence of diamond traits
Summary:
When loading a class, the number of methods in a class was initially estimated by `loadUsedTraits` without accounting for duplicated methods due to diamond trait inheritance.  This could incorrectly trigger the error "Class has more than 65.535 methods" (see the repro in `donotoverestimate.php` below).

This diff refines the estimation of the number of methods returned by `loadUsedTraits` so that methods imported multiple times due to diamond inheritance are counted only once.

Reviewed By: paulbiss

Differential Revision: D67594513

fbshipit-source-id: 194266a50dfffba2b9cd4e1bebb01fd7f9e07363