folly async io_uring: bump EVB start time on IoSqeBase completion
Summary:
`EventHandler::libeventCallback()` calls `EventBase::bumpHandlingTime()` which sets the `startTime_` of an EVB loop iteration if it wasn't already set. If no `EventHandler`s are run, then `startTime_` is bumped just before running loop callbacks.

The io_uring based `IoUringBackend` and `AsyncIoUringSocket` do not use `EventHandler`. Thus the `startTime_` of an EVB loop iteration is not being set correctly upon handling the first piece of I/O work via the various `IoSqeBase` types.

Fix this by adding the EventBase to that an IoSqeBase is tied to, and call `EventBase::bumpHandlingTime()` when completing an IoSqeBase i.e. invoking its callback.

Reviewed By: yfeldblum

Differential Revision: D68537153

fbshipit-source-id: 9e991213d69ffa32778e06a0fb447b0514bebba5