error if a custom PACKAGES.toml file cannot be open
Summary:
If a `PACKAGES.toml` file cannot be opened at startup then the system silently returns an empty package specification.  This can be confusing, especially when testing the build with custom package files.

This diff causes HHVM to emit an error if a custom package file is specified but cannot be opened (this is always an error).

Reviewed By: mdko

Differential Revision: D67452014

fbshipit-source-id: 94cfaecb980054be7f7032b994d6b2a5571c2dea