Eliminate false positives on unsurfaced lambda
Summary:
Lints are check on TASTs both under normal and dynamic assumptions.

Unsurfaced exception lint was producing false positives because the `dynamic` run had all lambdas replaced with `Omitted` expressions and the lint rule relies on finding references to the exception binding.

This diff opts the lint into sound linting which means that we require an agreement between the normal and dynamic pass. Since no lint is produced in the normal pass, the false positives are prevented.

Reviewed By: andrewjkennedy

Differential Revision: D68631579

fbshipit-source-id: 38efae1606fdc37250cb33e41da566183b2962f4