Record the function and method span in the typing env
Summary:
The `callable_pos` in the typing env currently records the position of the function or method _name_ which is used for error reporting on the wellformedness of functions (e.g. missing return hint).

This diff adds another field that so we record the span of function or method as part of the extended error context

Reviewed By: andrewjkennedy

Differential Revision: D68623863

fbshipit-source-id: db5a7197698b66950e0e89ac90e798f1cab01ca4