Make File_content.position abstract
Summary: File_content has a `position` type whose line and numbers are one-based. Users of the type need to be aware of that to use it properly. In this diff I'm making the type abstract so that users don't need to know the implementation. API will be provided for both zero-based and one-based versions of common operations.

Reviewed By: mheiber

Differential Revision: D68495085

fbshipit-source-id: efac671f73ad7d1fe2266eaa138c8a285dbc2c8f