Stop dirsyncing fixtures again
Summary:
dirsync was temporarily enabled for fixtures in D68628003 so that ShipIt could be switched to xplat without causing massive code churn on GitHub. Now that xplat is the source of truth for both ShipIt and fbcode_builder, disable dirsync again, removing the fbcode copy of fixtures. Among other things, this fixes a regression in developer experience, particularly with source control operations.

Enable and fix `build_fixtures` and `fixture_test` in xplat.

Reviewed By: yoney

Differential Revision: D68730277

fbshipit-source-id: dc77903caff7558cd95fb27bcbc1160284c8efcf