Revert D68464053: Multisect successfully blamed "D68464053: [thrift][patch] In DiffVisitor, if field exists in source, always ensure first" for one test failure
Summary:
This diff reverts D68464053
D68464053: [thrift][patch] In DiffVisitor, if field exists in source, always ensure first by Mizuchi causes the following test failure:

Tests affected:
- [cogwheel:cogwheel_controller_sizing#test_create_finalize](https://www.internalfb.com/intern/test/281475113955782/)

Here's the Multisect link:
https://www.internalfb.com/multisect/19963119
Here are the tasks that are relevant to this breakage:
T191387227: Some tests unhealthy for capacity_orchestration

The backout may land if someone accepts it.

If this diff has been generated in error, you can Commandeer and Abandon it.

Reviewed By: Mizuchi

Differential Revision: D68790038

fbshipit-source-id: dc37d0cca7b4c6768d1a883bd48a17ed380f70d9