Update comment in Code_actions_services
Summary:
We had a comment pointing to a potential bug, but there is no bug afaict, at least from the user POV and from the POV of our tests (which seem to be accurete).

Update the comment, with links to videos and tests.

It would be really nice not to have a WTF at all, but I once put ~3 hours into rooting it out and gave up.

Reviewed By: CatherineGasnier

Differential Revision: D68558573

fbshipit-source-id: fdb8c5dee6a44820f09a3dcf8210706751b6831d