introduce support for shape creation
Summary:
This diff introduces shape creation syntax, so that one can write:

```
$person = ETDemoObject::init(shape(
        'first_name' => 'John',
        'last_name' => 'Doe',
      ));
```

instead of

```
 $person = ETDemoObject::create($builder ==> {
        $builder->first_name = 'John';
        $builder->last_name = 'Doe';
      });
```

This has a few benefits:
* more succint to write
* proper typing (the latter pattern doesn't error on missing fields, for which one currently has to write lint rules)
* much more efficient runtime as it avoids additional closures and many small update operations

This also makes shape types useful in expression trees, for example the following is now valid:

```
ExampleDsl`{
    $f = (shape('x' => ExampleInt) $shape) ==> 3;
    $f(shape('x' => 2));
  }`;
```

Note that the shape type can be aliased as well.

This diff does not include a way to read from- or write to a shape (the proposal for that is in progress). However, this proposal is useful without it as we primarily intend to use shapes as argument to APIs that are code-genned, in which case a macro that casts the shape to an untyped map is sufficient.

Reviewed By: dlreeves

Differential Revision: D66576594

fbshipit-source-id: 451044e56d284e070b1deb8b5e7b1bedbf45ed52