Make isinstance checks pass on thrift.py3.GeneratedError
Summary:
Tl;dr: Basically the same diff as D68765988 but for exceptions.

When we're in auto-migrate mode, we want instance checks against `thrift.py3.GeneratedError` to still pass (as there's a lot of user code that checks `isinstance(my_struct, thrift.py3.GeneratedError)` to do some logic that is for thrift-py3 and otherwise do logic that is for thrift-py-deprecated.

This makes it such that for thrift-python exceptions, `._fbthrift_auto_migrate_enabled()` is checked, which is a bool that gets set to true when auto-migrate is enabled, so in normal build mode, `isinstance(python_struct, thrift.py3.GeneratedError)` returns `False`, but in auto-migrate mode returns `True`

Reviewed By: ahilger

Differential Revision: D68861603

fbshipit-source-id: 5cf7d001e2d43a36ec36575bbf6d64d10372606c