Set breakpoints on multiple non-contiguous bytecode blocks per line
Summary:
One line in source code may be translated to multiple blocks of bytecodes. For example, the local iterator optimization in HackC creates two copies of the foreach loop body. We need to set breakpoints on each of these blocks as long as they are not overlapping or contiguous.
Unfortunately, there are some cases where this may lead to spurious breakpoints being set. For example, multiline call expressions produce two bytecode blocks for the same line that are not contiguous, so there will be two breakpoints set.

Reviewed By: paulbiss

Differential Revision: D68462841

fbshipit-source-id: 824e55946b9a18babad658b1f926f5e26721a46f