Implement rendering for partials
Summary:
This diff implements rendering for partial blocks and statements as defined in the spec.

To achieve this, we have an *implementation-defined* `partial_definition` class that is exposed to Whisker using a `native_handle<>`. I especially enjoy the demonstrate of the capabilities of partials in the [collatz conjecture](https://en.wikipedia.org/wiki/Collatz_conjecture) example.

Differential Revision: D68687668

fbshipit-source-id: c85e044109ec70ad8260056a23c7fb65332ffc5f