Remove decls from oxidized decl parser state, collect as output of smart constructors
Summary:
The direct_decl_smart_constructors expect to be able to mutate the `Impl` state struct

https://www.internalfb.com/code/fbsource/[d7429589c3c9c8fa32f9b9352ca4733de3a25bb4]/fbcode/hphp/hack/src/decl/direct_decl_smart_constructors.rs?lines=2883-2884

but also have it resilient in the case of the parser backtracking.

https://www.internalfb.com/code/fbsource/[5871d8bd76f4e8d86498954b34ecc971fa2a3a48]/third-party/tp2/rust/1.83.0%2Bllvm.17/rustc-1.83.0-src/library/alloc/src/rc.rs?lines=1795-1797

Since the decls are stored in the state, this means they get cloned, which is cheap enough for arena-allocated oxidized-by-ref values but prohibitively expensive for owned oxidized values. However, we don't really need to store decls in the state -- it's much clearer if the smart constructors for top level declaration nodes actually return the nodes they represent instead of recording being a side-effect (see one of the updated comments)

In the future, we should be able to shift the cloning of the state up to the parser itself where the backtracking happens.

adapted from D47649144

Differential Revision: D68933279

fbshipit-source-id: 165f8492a981dab7de3fb21587a0eba2287dd42c