Implement CompareThreeWay using EqualTo and LessThan
Summary:
Implement three-way comparison and use it in `StructLessThan` as part of the effort to improve the performance of comparing two Thrift structs. For now, `CompareThreeWay` simply delegates to `EqualTo` and `LessThan`. This will serve as the default fallback implementation when the input types are not three-way comparable. As a result, there are no significant changes in the performance of `StructOpLess`.

#buildall

Reviewed By: Mizuchi

Differential Revision: D68541164

fbshipit-source-id: ea7a50de16c648be3ea2e3f5ec9ba5f92ca0dece