Fix checking of enum class constants under pessimisation
Summary:
When checking typing of enum class constants, we need to pessimise the expected type in `MemberOf`, because expressions of like-type are permitted on the right-hand-side. But we don't want this pessimisation to leak into other aspects of enum class typing, such as the parameters to `BuiltinEnumClass`. The pessimisation of the API of these classes (e.g. `getValues`) ensures safety here, just as with user use of generics.

We take the opportunity to do a bit of clean up and refactoring of the `MemberOf` processing in the direct-decl parser.

Differential Revision: D68954263

fbshipit-source-id: 5f4559d13b9b6b1619f4a1bc8bf8ee8e6ce8834f