implement key uniqueness validator
Summary:
A validator for map const values that flags when there are duplicate keys. This is a problem because the `const map` will have different values in different languages. For example, in thrift-cpp2 and thrift-py3, the first key-value pair is retained and others are discarded. In thrift-python, the last key-value pair overwrites the previous one(s).

```
[WARNING:configerator/structs/admarket/lib/adeventtraits/adevent_traits.thrift:4016] Duplicate key in map literal: `instagram_profile_follow`
```

In next diff, I will make the key error printout better for enums.

Reviewed By: vitaut

Differential Revision: D68657586

fbshipit-source-id: 9811d2d3e1dfe5bfd7721ce4df0f67265a4b9615