Add tests for oxidized decl parser
Summary: Similar to D68767994, renames a handful of APIs that produce arena-allocated decls, just enough to make Hh_single_decl run over test/decl. This gives us a correctness test of the oxidized DDP to make sure `_smart_constructors` and `_smart_constructors_obr` remain in sync. I left the `--decl-parse` flag in there for ease of testing; it is unused

Differential Revision: D68934722

fbshipit-source-id: 9bfa13735458835d9c5ce4e9989797e2fc392e44