Patch a few differences, enable the test/typecheck:verify_oxidized target
Summary:
Cleans up a few discrepancies between oxidized and oxidized-by-ref DDPs. These are mostly relevant to error cases so when we delete OBR we can revert this change.

Alternatively, we accept the differences as they are now and roll with it, depending on if www is clean of delatas. **RFC**

 ---
The only changes I explicitly accepted are related to conditional declaration of functions and classes; the difference stems from the base of this stack D68933279, OBR statefully records each decl, whereas O does a simple collection of the children of script_declaration, and if statements are discarded by their smart constructors.

Differential Revision: D69036626

fbshipit-source-id: c72c48ce904e981944b726fc71451a23d574db26