Fix bug where shape fields where not checked in open shapes
Summary:
We have Tshape.check_validity_of_field or something. This is called in a couple of places, like typing_typedef, after matching a hint on Hshape.
This is brittle because sometimes the Hshape is wrapped in HH\supportdyn.
I'm moving the check into Typing_wellformedness_check to make it more robust.
I had to wire the env through there, so a lot of the lines in this diff are due to that.

Reviewed By: andrewjkennedy

Differential Revision: D68564144

fbshipit-source-id: bffdf0454fe61d489b527902b63d3f5b77c26113