use same xid in propagation
Summary:
Now that we're passing in the service interface to the context prop handler, we can use the string as the correct service name without having to do any additional transformations.

This will help us consolidate the logic between the flows to check and the propagation handlers and resolve all of the current discrepancies by using the same string.

Reviewed By: zhonglowu2

Differential Revision: D68980419

fbshipit-source-id: 135a82da1afbdffbcffb330cb4d7c28643029374