Adjust backtrace naming to account for import statements
Summary:
Currently, we print `<partial-name> @ <file>` for partials and `<file>`  for macros.

With the introduction of import statements, we need an additional jump reason.

This diff adjusts the output so that macros become `<macro> @ <file>` and imports take on `<import> @ <file>`. The bare `<file>` is used for the root source file only.

Reviewed By: yoney

Differential Revision: D69273909

fbshipit-source-id: f137722414e56239e5787245e751a6de9791d958