Fix TSAN issue related to LOG_EVERY_N
Summary:
LOG_EVERY_N uses two integers to keep track of how many times the  line has been hit.  If the code is run from multiple threads this will cause TSAN errors.  Implement our own version that uses an atomic.

This will be a bit more expensive due to the `%` operator, but we can afford that in this code.

Reviewed By: aditya-jalan, sdunster

Differential Revision: D69489221

fbshipit-source-id: f2c191bdf11f3acf8027f4d6fdbbd46b1ef76ce4