Intersect left with predicate
Summary: When we partition a type, we may use more information about the type than may be known in the local context. This is true for case types (as demonstrated by the test) and may also be true if we use information like `__Sealed` to partition class types. This means that although we may judge a type to be completely left of a predicate, the user code may still expect an intersection to be produced.

Reviewed By: dlreeves

Differential Revision: D68470060

fbshipit-source-id: 02787b65994c2013ed0839b925a824f1ddc98c89