Add ~safe_for_are_disjoint flag to typing_case_types
Summary:
NOTE: This change is required for D67359646 which create class predicates from `is SomeClass` conditions. It is a separate diff for ease of review.

There's an issue with this where sometimes the datatype for `IFoo` and `not IFoo` are not recognized to be disjoint if `IFoo` has `require extends`.

This manifests as real problems when using this to split types for refinements. This flag allows turning off certain "smart handling" of sealed attributes and requirements.

This flag will be turned on in a subsequent diff b/c it cannot be turned on w/o doing type splits on class hints in is-expressions.

Reviewed By: dlreeves

Differential Revision: D68470061

fbshipit-source-id: 8a246f7f6bec17707dfb281223722439653263f0