Bypass subtyping to obtain "assumptions" from type splitting for GADT-style refinement
Summary:
Based on dlreeves's D60312834 -- a temporary shortcut until we've refactored subtyping/environment to store refinements more broadly

Previous to this diff, given a `if ($foo is Bar)` when we can turn `Bar` into a predicate, we produce a TypeSwitch constraint type relating `$foo` to two new type variables that represent the type of `$foo` in the true and false branches. Subtyping of this constraint type results in adding bounds to those type variables based on the the type split of the type of `$foo` using the predicate.

This diff skips the constraint type and subtyping and does the type split directly. This allows us to produce and use subtype propositions created as a result of the splitting process. This is useful when we can conclude something new on the true side of the split—for example in:

```
final class MyClass implements IMyInterface<int> {}

function foo<T>(IMyInterface<T> $x): T {
  if ($x is MyClass) {
    // we can conclude in this branch that int < T
  }
}
```

Previous to this diff we would only conclude this via a special refinement in typing.ml (https://fburl.com/code/b28wvua5) but we'd like to do this more generally as part of type splitting. Subtyping doesn't currently let us export updates to the environment so this diff allows us to export updates to the environment as part of type splitting by bypassing the subtyping with the constraint type when not testing a type variable.

No new expressiveness is gained by this diff, but it opens the door for D64611956

Reviewed By: dlreeves

Differential Revision: D64499771

fbshipit-source-id: 5345ac8016bd9f1d9d72765d866c402f94367302