Typing_intersection.intersect_lists - intersect neg types last
Summary:
NOTE: This change is required for D67359646 which create class predicates from `is SomeClass` conditions. It is a separate diff for ease of review.

Here's an example that I ran into:

Intersect `(!G & P)` and `?F` where `F < P` and `G < P`; Previously this would happen by:

* `?F` gets normalized to `null|F`
* We `intersect_lists` `[(null|F)]` `[!G ; P]`
  * We `intersect_ty_tyl` `(null|F)` `[!G ; P]`
    * We `try_intersect` `!G` and `(null | F)`
       * This becomes `(null | (!G & F))` which is considered successfully intersected because it's a union at the top-level
    * We `try_intersect` `(null | (!G & F))` `P`
       * This becomes `!G & F` but is considered unsuccessful because it's an intersection at the top level
    * Because the last `try_intersect` was an intersection, we throw it away and produce the unsimplified intersection `P & (null | (!G & F))`

This final type is cumbersome.

If we instead intersect the negative last, the first `try_intersect` of `(null|F)` and `P` produces `Some F` which we'll then intersect with `!G`.

We recognize that intersection can be order-dependent (see `~hint_first` argument in the codebase) but this hopefully avoids needing to fiddle with this at the API layer sometimes.

Pros: Assuming this is a good heurstic, it means less fiddling with order at the API layer. e.g. instead of D68470062 (Unfortunately, this change doesn't remove the need for `~hint_first` -- I tried always doing hint first after this change and got ~80 errors; I might fiddle with it more though)
Cons: It reduces the ability to fiddle at the API layer since we'll re-order here in the implementation.

Reviewed By: dlreeves

Differential Revision: D69200797

fbshipit-source-id: 816fa268060fd8cca4a35cf7e8b1ba211e1e4618