Implement CompareThreeWay using EqualTo and LessThan
Summary:
Implement three-way comparison and use it in `StructLessThan` as part of the effort to improve the performance of comparing two Thrift structs. For now, `CompareThreeWay` simply delegates to `EqualTo` and `LessThan`. This will serve as the default fallback implementation when the input types are not three-way comparable. As a result, there are no significant changes in the performance of `StructOpLess`.

For now, keep only 1 Tag in the template parameter of `CompareThreeWay` since we only use it to compare the exact same field in two Thrift structs.

Additionally, add a `LessThanImpl` template parameter to `CompareThreeWay` to ensure that types that do not have `operator<` defined can rely on it for less than comparison.

The default handler for adapted types is included in this diff to allow tests to pass. A more elaborate implementation to handle adapted types will be included in the next diff.

This is the second attempt of D68541164 after it was reverted due to test failures.

#buildall
#buildmore

Reviewed By: Mizuchi

Differential Revision: D69239177

fbshipit-source-id: 4b6e584940b244cd3a372831cb06a8d8e81e880d