Implement a live variable analysis for expression trees
Summary:
Implement a live variable analysis for expression trees.

Macro-style nested expression trees will need to use this to detect which variables are used w/o first being assigned so that we know which variables need to be macroed in.

See D66956540 for more context

NB, `lambda_captures.rs` looks like it's trying to compute this, but it's wrong (it goes forward rather than backward). We could think about replacing that with this, but for now, we're just working on the part of Hack that can appear in ETs.

Reviewed By: mheiber

Differential Revision: D66703669

fbshipit-source-id: e0af342513589f35b69750c55b40efc293565a7d