Disable JIT entirely when JIT is folding functions
Summary:
When JIT is folding functions, we rely on enterVMAtFunc() to enter dispatch(),
which generally does not reenter JIT.

However, this is broken. If the interpreted function threw an exception, the
enterVM() logic would then attempt to handle the exception and potentially
reenter using enterVMAtCurPC(), which lacks a similar check.

It is safer to just disable RID().getJit() during folding.

Also fix a bug in handleResume() that would try to use an existing translation
if JIT is disabled.

Allows the next diff to kill dispatch().

Reviewed By: ricklavoie

Differential Revision: D69623648

fbshipit-source-id: 43ed7c362d496da9ea2081b209011181357e960f