Fix locking of const objects while unwinding
Summary:
The previous logic was incorrect -- it would not lock objects if their
constructor was called from an interpreted FCallCtor via dispatchBB(),
since that would enter the unwinder via unwind-itanium even if nothing was
jitted. The unwind-itanium logic uses the fpToUnwind parameter to unwind
one frame at a time, which was conflated with "the caller was jitted".

Reviewed By: ricklavoie

Differential Revision: D69815680

fbshipit-source-id: c97084adba96ae2f1b1bfdff9cda52d54a136b17